requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
brotli==1.0.9
    # via flask-compress
click==8.1.3
    # via flask
dash==2.5.1
    # via -r requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
dnspython==2.2.1
    # via pymongo
flask==2.1.2
    # via
    #   dash
    #   flask-compress
flask-compress==1.12
    # via dash
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via jinja2
numpy==1.23.1
    # via pandas
pandas==1.4.3
    # via -r requirements.in
plotly==5.9.0
    # via dash
pymongo[srv]==4.1.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-decouple==3.6
    # via -r requirements.in
pytz==2022.1
    # via pandas
six==1.16.0
    # via python-dateutil
tenacity==8.0.1
    # via plotly
werkzeug==2.1.2
    # via flask
zipp==3.8.0
    # via importlib-metadata
